Fix for directly imported JSON configs being deletable from original source (Thanks TikiPrime)

+++ SynthEBD/GUI_Aux/FileDialogs.cs
@@ -15,7 +15,7 @@ namespace SynthEBD
             {
                 try
                 {
-                    if (!string.IsNullOrWhiteSpace(path))
+                    if (!string.IsNullOrWhiteSpace(path) && System.IO.File.Exists(path))
                     {
                         System.IO.File.Delete(path);
                     }
